pyproject.toml
[project]
name = "fastapi-infisical-project"
version = "0.1.0"
description = "FastAPI project with Infisical, PostgreSQL, and Redis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastapi", "uvicorn", "psycopg2-binary"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "pyright>=1.1.350",
]

[tool.hatch.build.targets.wheel]
include = ["main.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"